Calendar: Update typing
I've been struggling to keep up with the various types, so I figured
it's time to put in the effort to clean them up.

In doing so, this allowed me to add an 'id' to each event. This will
make it much easier to edit and remove events (coming later)