Events: Fetch from DB on load
Does the thing! Fetches all current events stored in the db upon page
load. This needs some tweaking (surprise, surprise, the dates seem to be
in the wrong timezone) but it works for the most part. Unfortunately,
`start` and `end` seem to be reserved keywords in Neon/Postgres, so I
had to rename the fields to `start_time` and `end_time`.

I set a 5 minute cache to start with, but this might change.

Finally, as the comment says, I plan on fetching the current month, +/-
1 month on page load.