Events: Save to database
Okay, this is a larger commit than I normally like, but I struggled a
bit so I want to commit while it's working. We're now storing and
loading events to the database!

I had hoped to make these server actions, like the original fetch was,
but Next never seemed to accept it. API calls it is.

Changes to be made:

1. Use an environment variable for the URL, instead of hardcoding the
   `localhost`
2. Caching
3. Return the id of the newly created events so we can remove the
   `numEvents` global
4. Wrap the get function in a helper that accomplishes the formatting
   for us
5. Alert the user if the DB api call fails so they can try again